+++ src/main/kotlin/com/alugames/models/FirstPlan.kt
@@ -1,21 +1,23 @@
 package com.alugames.models
 
+import java.math.BigDecimal
+
 class FirstPlan(
     type: String = "PRATA",
     val gamesIncluded: Int,
     val fees: Double,
-    private val discountPercentage: Double,
+    private val discountPercentage: BigDecimal,
 ): Plan(type) {
 
-    override fun rentPrice(rent: GameRent): Double? {
+    override fun rentPrice(rent: GameRent): BigDecimal? {
         val gamesRented = rent.monthGamesRented(rent.periodRent.initialDate.monthValue).size + 1
         val price = super.rentPrice(rent)
 
         return if (gamesRented <= gamesIncluded) {
-            0.0
-        } else if(rent.gamer.media > 8) {
+            BigDecimal(0)
+        } else if(rent.gamer.media > BigDecimal(8)) {
             val discount = price?.times(discountPercentage)
-            price?.minus(discount ?: 0.0)
+            price?.minus(discount ?: BigDecimal(0))
         } else {
             price
         }
