+++ src/main/kotlin/com/alugames/models/Recommended.kt
@@ -1,6 +1,8 @@
 package com.alugames.models
 
+import java.math.BigDecimal
+
 interface Recommended {
-    val media: Double
+    val media: BigDecimal
     fun recommend(note: Int)
 }
